+++ src/prelude.rs
@@ -1,0 +1,3 @@
+#![allow(clippy::pub_use)]
+pub use crate::merkle_bit::*;
+pub use crate::traits::*;