+++ src/tree/tree_data.rs
@@ -77,7 +77,7 @@ impl Encode for TreeData {
 impl Encode for TreeData {
     #[inline]
     fn encode(&self) -> BinaryMerkleTreeResult<Vec<u8>> {
-        Ok(serde_yaml::to_vec(&self)?)
+        Ok(Vec::from(serde_yaml::to_string(&self)?))
     }
 }