+++ src/tree/tree_leaf.rs
@@ -121,7 +121,7 @@ impl<const N: usize> Encode for TreeLeaf<N> {
 impl<const N: usize> Encode for TreeLeaf<N> {
     #[inline]
     fn encode(&self) -> BinaryMerkleTreeResult<Vec<u8>> {
-        Ok(serde_yaml::to_vec(&self)?)
+        Ok(Vec::from(serde_yaml::to_string(&self)?))
     }
 }