+++ src/tree_db/hashmap.rs
@@ -2,7 +2,7 @@ use crate::Array;
 use std::collections::hash_map::HashMap;
 use std::path::Path;
 
-use crate::traits::{Database, Exception};
+use crate::traits::{Database, MerkleBitError};
 use crate::tree::tree_node::TreeNode;
 
 /// A database consisting of a `HashMap`.
@@ -19,6 +19,7 @@ impl<const N: usize> HashDB<N> {
         Self { map }
     }
 
+    #[allow(clippy::missing_const_for_fn)]
     #[inline]
     #[must_use]
     /// Decomposes the `HashDB` into its underlying `HashMap`.
@@ -28,15 +29,13 @@ impl<const N: usize> HashDB<N> {
 }
 
 impl<const N: usize> Database<N, TreeNode<N>> for HashDB<N> {
-    type EntryType = (Array<N>, Vec<u8>);
-
     #[inline]
-    fn open(_path: &Path) -> Result<Self, Exception> {
+    fn open(_path: &Path) -> Result<Self, MerkleBitError> {
         Ok(Self::new(HashMap::new()))
     }
 
     #[inline]
-    fn get_node(&self, key: Array<N>) -> Result<Option<TreeNode<N>>, Exception> {
+    fn get_node(&self, key: Array<N>) -> Result<Option<TreeNode<N>>, MerkleBitError> {
         self.map.get(&key).map_or(Ok(None), |m| {
             let node = m.clone();
             Ok(Some(node))
@@ -44,19 +43,19 @@ impl<const N: usize> Database<N, TreeNode<N>> for HashDB<N> {
     }
 
     #[inline]
-    fn insert(&mut self, key: Array<N>, value: TreeNode<N>) -> Result<(), Exception> {
+    fn insert(&mut self, key: Array<N>, value: TreeNode<N>) -> Result<(), MerkleBitError> {
         self.map.insert(key, value);
         Ok(())
     }
 
     #[inline]
-    fn remove(&mut self, key: &Array<N>) -> Result<(), Exception> {
+    fn remove(&mut self, key: &Array<N>) -> Result<(), MerkleBitError> {
         self.map.remove(key);
         Ok(())
     }
 
     #[inline]
-    fn batch_write(&mut self) -> Result<(), Exception> {
+    fn batch_write(&mut self) -> Result<(), MerkleBitError> {
         Ok(())
     }
 }